changes from mentor

+++ ServiceApi/InnoClinic.ServiceApi.BL/Dto/Service/CreateServiceRequest.cs
@@ -4,7 +4,7 @@ public class CreateServiceRequest
 {
     public Guid CategoryId { get; init; }
     public string Name { get; init; }
-    public float Price { get; init; }
+    public decimal Price { get; init; }
     public Guid SpecializationId { get; init; }
     public bool IsActive { get; init; }
 }
