+++ ServiceApi/InnoClinic.ServiceApi.BL/Dto/Service/ShowServiceResponse.cs
@@ -5,7 +5,7 @@ public class ShowServiceResponse
     public Guid Id { get; init; }
     public Guid CategoryId { get; init; }
     public string Name { get; init; }
-    public float Price { get; init; }
+    public decimal Price { get; init; }
     public Guid SpecializationId { get; init; }
     public bool IsActive { get; init; }
 }
